Don't switch environment before release build
Simply put, this doesn't work. This code already runs in Mix, and Mix
has already loaded the appropriate env config, and changes to the
active environment at this stage have no effect on that config.

The correct way to deploy a production release must be:

    $ MIX_ENV=prod mix knarr.deploy <name>

... which doesn't prevent the deployment of releases built with
MIX_ENV=dev, and it's also closer to how `mix release` is meant to be
used anyway.